==> 05-results.Rmd <==
# Results


```{r}
library(tidyr)
library(dplyr)
library(ggplot2)
library(heatmaply)
library(scales)
library(directlabels)
library(knitr)
#devtools::install_github("UrbanInstitute/urbnmapr")
library(urbnmapr) 
library(ggplot2)  
library(dplyr)   
library(stringr) 
library(gridExtra)


library(lubridate)
library(choroplethr)
library(choroplethrMaps)
library(Hmisc)
library(ggthemes)
library(viridis)
library(geofacet)

```

```{r}
load(file="data/covid.RData")
load(file="data/trips.RData")
load(file="data/covid_cases_v5.Rda")
air_data <- read.csv("data/BTS Domestic Air Travel 2020.csv")
```

## Covid-19 in the United States 

In order to compare the mobility of people against the transmission of Covid-19 we must first investigate how the virus spread throughout the United States. 

Main approach we had taken for this project was exploring Covid data at macro level first, then subsequently moving into micro perspective. 
Can we find any insight from the state level, big picture data? 


```{r}
ggplot(covid_state_pct, aes(x=Date, y=Positive_pct)) + 
  geom_line(color="steelblue", size=1) +
  theme_bw() +
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.x=element_blank(),
        axis.ticks.y=element_blank()) +
  labs(title = "COVID RATES IN US STATES") +
  ylab("Covid positive rates (%)") +
  facet_geo(~State, grid="us_state_grid2")
```

R package, ‘geofacet’, doing a great job here to display time-series of each state visualized on a US map. This plot tells us the overall Covid trend across the states. No surprise here to find that Covid rates accelerated around the last summer, then slowed down in most states. We also observe that some northern states (e.g Vermont) and geographically isolated regions (e.g. Hawaii) have Covid rates plateaued at relatively low levels. But every state is different in size, we need more analysis to see any relations with population size, mobility, traffic and so on. Let’s move to the next plot. 




```{r}
ggplot(covid_state_pct_monthly, aes(x=month,y=positive_pct_monthly))+
  geom_point() +
  geom_smooth() +
  facet_grid(cols=vars(population.cat)) +
  scale_fill_viridis() +
  scale_y_continuous(labels = comma) +
  theme_bw() +
  theme(legend.position = "None") +
  labs(title = "COVID RATES IN US STATES: DIFFERENT SIZE GROUP") +
  ylab("Covid positive rates (%)") +
  xlab("Month") 
```

Naturally, we get curious to find out whether Covid trends differ across different population size in each state. We divided all US counties and states into three sub-categories (‘large’, ‘medium’, ‘small’), and plotted three separate facet grids for each size configuration. 

For this graph, ‘geom_point()’ was used to display multiple time-series in one graph. We can clearly see that ‘small’ population states show wider variance of Covid rates, especially going through multiple waves of Covid spread in the latter part of the year. And interestingly enough, most states with higher than 10% Covid rates belong to this ‘small’ category. 

Can we hypothesize that a smaller population leads to higher Covid rates? Does this scatter plot have enough telling? Not so fast ! Soon after we visualized the average trendline using ‘geom_smooth()’ function, we had come to find that overall Covid trend remain similar to each other across three different size categories afterall. 



```{r}
df_bar <- covid_state_pct_monthly %>% 
  group_by(population.cat, month) %>% 
  dplyr::summarise(avg_positive_pct = mean(positive_pct_monthly))

ggplot(df_bar, aes(x=month, y=avg_positive_pct, fill=population.cat)) +
  geom_bar(stat="identity") +
  labs(title = "COVID RATES IN US STATES: DIFFERENT SIZE GROUP") +
  ylab("Covid positive rates (%)") +
  xlab("Month") +
  theme_bw() +
  theme(legend.title = element_blank(),
        legend.position = "bottom")
```

Covid rates plateaued around 10% across the states, can we observe the same pattern during the early spike in Covid cases in the US? Let’s have a closer look at Covid trend since the first outbreak in March 2020. For this, we chose a stacked bar chart to compare case numbers across categorical variables (population size). 

Do you also see the large states dominated the early Covid spikes, say for the first 100 days or so? Perhaps, is this telling that there might be some connection between Covid outbreak and flight traffic in-and-out of the major airport hubs? Or, is this due to the density of the region and mobility of people residing there? We will explore this further using other dataset as we progress our project further joining additional data.

In the meantime, as many of us are living near New York City, we thought it is a good practice to carefully check what happened in the Tri-State area and its major counties in early 2020.




```{r}
ny_county = c("New York County", "Kings County", "Bronx County", "Richmond County", "Queens County")
nj_county = c("Bergen County", "Hudson County")
ct_county = c("Fairfield County", "New Haven County")

ny_FIPS = c(36061, 36047, 36005, 36085, 36081)
nj_FIPS = c(34003, 34017)
ct_FIPS = c(9001, 9009)

first100 <- covid_county_net %>% 
  filter(County.FIPS %in% c(36061, 34003, 9001)) %>% 
  filter(Date > "2020-03-01" & Date < "2020-06-30")

ggplot(first100, aes(x=Date, y=Positive.change, color=County.Name)) +
  geom_line() +
  geom_smooth() +
  labs(title = "TRI-STATE AREA COUNTIES", subtitle = "First 100 days") +
  ylab("Covid positive cases (Net change)") +
  xlab("") +
  scale_fill_viridis() +
  theme_bw() +
  theme(legend.title = element_blank(),
        legend.position = "bottom")
```

We filtered our daily Covid data only until the end of June last year, and we selected only one major county from each NY, NJ and CT. 

We need to admit that choosing the first 100 days, and only three counties from a vast list of Tri-state counties are rather subjective decision. Other than 100 day being symbolic, we anecdotally remember the first outbreak and first wave of Covid spikes happened before the summer. In fact, the above time-series graph shows that the first peak of daily increase happened before May 2020. Plus, given its geographical proximity to New York City, Bergen and Fairfield counties are reasonable candidates for monitoring early Covid spread around neighboring regions.

Can you also see an early spike happened first in New York, then followed by Bergen and Fairfield counties? There were a few weeks of time-lag until Bergen and Fairfield numbers did catch up with New York’s. 

Next question follows. What if we compare the entire county data in the Tri-state area, not just previously chosen three counties? By the way, to answer this question we need to remind ourselves that each county differs widely in size. So, to make a fair comparison, we will use Covid rates (in percentage) instead of case numbers.




```{r}
tri_state <- covid_county_pct %>% 
  filter(State %in% c("NY", "NJ", "CT")) %>% 
  filter(Date > "2020-03-01" & Date < "2020-06-30")
  
ggplot(tri_state, aes(x=Date, y=Positive, color=population.cat, shape=population.cat, alpha=0.5)) +
  geom_point() +
  geom_smooth() +
  labs(title = "COVID RATES IN TRI-STATE AREA", subtitle = "First 100 days") +
  ylab("Covid positive rates (%)") +
  xlab("") +
  scale_fill_viridis() +
  scale_alpha(guide = 'none') +
  theme_bw() +
  theme(legend.title = element_blank(),
        legend.position = "bottom")
```

Do you see the early Covid surge was mainly driven by big, large counties? Somehow, this finding reaffirms our previous observations that large states with big populations dominated the early Covid spikes before converging to each other in later stages. Again, we wanted to find out the reason why this happened, is it because of traffic? Or people in the densely populated area wanted to move around more actively despite Covid outbreaks? Let’s find out. 





## Mobility and Covid-19


```{r}
test <- trips.county
test$perSAH <- test$Population.Staying.at.Home/ (test$Population.Staying.at.Home + test$Population.Not.Staying.at.Home)
test$perNSAH <- test$Population.Not.Staying.at.Home/ (test$Population.Not.Staying.at.Home + test$Population.Staying.at.Home)
test$Total_pop <- test$Population.Not.Staying.at.Home + test$Population.Staying.at.Home

```

```{r}
state <- trips.state
test$perSAH <- test$Population.Staying.at.Home/ (test$Population.Staying.at.Home + test$Population.Not.Staying.at.Home)
test$perNSAH <- test$Population.Not.Staying.at.Home/ (test$Population.Not.Staying.at.Home + test$Population.Staying.at.Home)
test$Total_pop <- test$Population.Not.Staying.at.Home + test$Population.Staying.at.Home
```



Now, using the mobility data from the Trips by Distance data from the U.S. D.O.T Bureau of Transportation Statistics, let us investigate the mobility of people within the U.S. and find the impact that travel may have had on the transmission of COVID-19.
In a similar fashion to the COVID-19 data, let us begin by looking at the macro and working our way to the micro. 

First, let us aggregate the data to find the average Percentage of People Not Staying at Home (perNSAH) during the pandemic: 




```{r}
#top bottom counties NSAH chart
avgNSAH <- test %>% 
  group_by(County,State) %>% 
  summarise(avgNSAH = mean(perNSAH)) %>% 
  arrange(desc(avgNSAH)) %>% drop_na() 

high <- apply(head(avgNSAH),2,as.character)
temp <- matrix(rep(".",9),ncol=3)
low <- apply(tail(avgNSAH),2,as.character)
kable(rbind(high,temp,low))

```


From this chart we can see a sample of the highest and lowest total averages of state’s mobility. This gives us a general idea of how the population of states traveled but, in order to produce significant results, we must investigate further the mobility of each state. 


To do this, let us take an aggregated look at how each state has been moving considering the distance of trips within each state. The heatmap below gives us an overall picture of how each state was comparatively mobile by distance traveled in miles during the pandemic. 




```{r}
#for heatmap 
test2 <- state[,-1] %>% 
  group_by(State) %>% 
  dplyr::summarise_all(.funs=sum)

test2 <- as.data.frame(test2)
rownames(test2) <- test2$State
test2 <- as.matrix(test2[,-(1:5)])
colnames(test2) <- c("<1","1-3","3-5","5-10","10-25","25-50","50-100","100-250","250-500","500+")
```

```{r fig.height=9}
heatmaply(
  percentize(test2),
  xlab="Distance in Miles Traveled",
  main="Heatmap of Miles Traveled by State"
)

```


By this plot we can see that the overall percentage of trips by state were higher in states like CA, TX, and NY, and lower in VT, WY, ND. With this we can see North Carolinians took fewer 500+ mile trips than people from Colorado. To note here, the 500+ mile bracket is unintentionally fixed to the left hand side of the graph however, this could also could be considered the most important travel bracket since every state’s diameter (aside from one direction in CA and TX) is no further than 500miles thus, any 500+ mile trip means the travelers crossed state lines. This necessarily means possible transmission of the virus to other states. 

Now, to get a better picture of mobility and compare this data against the concentrations of COVID-19, let us visualize the concentrations of infections and travel across the United States. 



```{r}
COV <- read.csv("data/covid_confirmed.csv", stringsAsFactors=FALSE)
names(COV)[1] <- "countyFIPS"  # Fix the name of first column. Why!?

```


```{r warning=F}
Covid <- pivot_longer(COV, cols=starts_with("X"), 
                     values_to="cases",
                     names_to=c("X","date_infected"),
                     names_sep="X") %>%                
  mutate(date_infected = as.Date(date_infected, format="%m.%d.%Y"),
         countyFIPS = str_pad(as.character(countyFIPS), 5, pad="0"))

states_sf <- get_urbn_map(map = "states", sf = TRUE)
counties_sf <- get_urbn_map(map = "counties", sf = TRUE)

counties_cov <- inner_join(counties_sf, group_by(Covid, countyFIPS) %>%
       summarise(cases=sum(cases)), by=c("county_fips"="countyFIPS"))

```


```{r warning=F}
p1 <- counties_cov %>%
  ggplot()+
  geom_sf(mapping = aes(fill = cases), color = NA) +
  geom_sf(data = states_sf, fill = NA, color = "black", size = 0.25) +
  coord_sf(datum = NA) +   
  scale_fill_gradient(name = "Cases", trans = "log", low='yellow', high='red',
                      na.value="white", breaks=c(0, max(counties_cov$cases)),
                      labels=c("1","1.2M")) +
  theme_bw() + theme(legend.position="bottom", panel.border = element_blank())+
  labs(title="Density of COVID-19 Cases")

```


```{r}
test$County.FIPS <- as.character(test$County.FIPS)
test <- test[!(is.na(test$Number.of.Trips)),]

test3 <- test %>% 
  group_by(County) %>% 
  summarise(perNSAH =mean(Number.of.Trips)) 

```


```{r warning=F}

counties_trips <- inner_join(counties_sf, test3,
                             by=c("county_name"="County"))

p2 <- counties_trips %>%
  ggplot()+
  geom_sf(mapping = aes(fill = perNSAH), color = NA) +
  geom_sf(data = states_sf, fill = NA, color = "black", size = 0.5) +
  coord_sf(datum = NA) +   
  scale_fill_gradient(name = "Trips",trans = "log", low='yellow', high='purple',
                      na.value="yellow", breaks=c(min(counties_trips$perNSAH),
                                                 max(counties_trips$perNSAH)),
                      labels=c("0","24.8M")) +
  theme_bw() + theme(legend.position="bottom", panel.border = element_blank())+
  labs(title="Density of Total Mobility")


```

```{r warning=FALSE }
grid.arrange(p1,p2,ncol=2)
```


This plot allows us to see where in the United States people are the most mobile and allows us to compare those regions to where the highest concentrations of coronavirus are located. Unsurprisingly, most of the mobility occurs around highly populated, urban areas and subsequently, most of the COVID cases are also typically centered around the same areas. So, the findings so far point in the direction of increased mobility being linked to increased positive cases. However, to confirm this assumption we must investigate further into the cases and mobility of certain regions because, to say comparing Los Angeles County (with a population of 10.4million people) to any other random county in the U.S. is like comparing apples to oranges. Thus, to arrive at a poignant destination we must first find a couple of apples.  

To do this, we will search for high and low mobility counties with similar stats (population, area, etc.) to compare and really be able to find and answer to the question of mobility affecting transmission. Accounting for density and population, the first matchup we have is: Pittsylvania County VA  v.s. Walla Walla, WA. 



```{r}
apple <- test[test$County.FIPS==c(51143,53071),]

temp <- apple %>%  group_by(State,County) %>% summarise(avgNSAH = mean(perNSAH))
temp$Land_area <- c("1,270 sq mi", "1,299 sq mi")
temp$Pop_Density <- c("65/sq mi","46/sq mi")
kable(temp)
```





```{r}

apple <- test[test$County.FIPS==c(51143,53071),]

a1 <- ggplot(apple, aes(x=Date,y=perNSAH,color=County))+
  geom_line()+
  labs(title="% Not Staying Home",ylab="Percent")+
  theme(legend.position="bottom")

```



```{r}
apple2 <- covid.county[covid.county$County.FIPS==c(51143,53071),]


a2 <- ggplot(apple2, aes(x=Date,y=Positive,color=County))+
  geom_line()+
  labs(title="Total Covid Cases",ylab="Total COVID Cases")+
  theme(legend.position="bottom")


grid.arrange(a1,a2,ncol=2)
```

As we can see from this plot there were more people Not staying at home in Pittsylvania than Walla Walla,  however, the total rates of covid remained similar. These counties are relatively small and not representative so, let’s look at another set to compare. Next up we have Brazoria County, TX v.s. Marin County,CA 


```{r}
apple <- test[test$County.FIPS==c(48039,06041),]

temp <- apple %>%  group_by(State,County) %>% summarise(avgNSAH = mean(perNSAH))
temp$Land_area <- c("828 sq mi", "1,609 sq mi")
temp$Pop_Density <- c("300/sq mi","267/sq mi")
kable(temp)

a1 <- ggplot(apple, aes(x=Date,y=perNSAH,color=County))+
  geom_line()+
  labs(title="% Not Staying Home",ylab="Percent")+
  theme(legend.position="bottom")
```

```{r}
apple2 <- covid.county[covid.county$County.FIPS==c(48039,06041),]


a2 <- ggplot(apple2, aes(x=Date,y=Positive,color=County))+
  geom_line()+
  labs(title="Total Covid Cases",ylab="Total COVID Cases")+
  theme(legend.position="bottom")


grid.arrange(a1,a2,ncol=2)

```


Here, more Brazorians went out than Marins and we can see a much higher rate of transmission in Texas. So, what does the comparison of these different counties tell us?
That the results are mixed. There is no definitive proof that the stay at home orders worked. There are possibly other variables that affect transmission more than the mobility of society. 



## Air Travel and Covid-19


```{r}
# Loading Air Travel Data and selecting important columns of interest

air_data <- air_data %>% 
  select(DEPARTURES_PERFORMED, PASSENGERS, DISTANCE, CARRIER_NAME, ORIGIN_CITY_NAME, ORIGIN_STATE_ABR, 
         ORIGIN_STATE_NM, DEST_CITY_NAME, DEST_STATE_ABR, DEST_STATE_NM, YEAR, MONTH)

#summary(air_data)
#head(air_data)
```

Continuing on the theme of mobility, while looking at Covid data through the air travel lens, the first obvious question is: what was air travel like in 2020 during this pandemic. The plot below aggregates the total number of passengers per month in 2020. It is clear that there is a significant drop in air travel across the United States at the beginning of the pandemic especially in the month of April. This makes sense as air travel in America was essentially grounded. It can be seen near the end of 2020 that air travel slowly started to pick back up, although not to previous levels before the pandemic.

```{r}
# Total Travel of Passengers in 2020 by Month
total_travel <- air_data %>%
  mutate(MONTH = month.abb[as.numeric(gsub("M","",MONTH))]) %>%
  mutate(MONTH = factor(MONTH, levels=c("Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct","Nov","Dec"))) %>%
  group_by(MONTH) %>%
  summarise(Total_Passengers = sum(PASSENGERS))

#head(total_travel)

ggplot(total_travel, aes(x = MONTH, y = Total_Passengers)) +
  geom_bar(stat = "identity") +
  scale_y_continuous(labels = scales::comma) +
  ggtitle("Total Passenger Airline Travel 2020 Jan-Dec")
```


Delving a little further into this result, can we correlate air travel to states, to spikes in Covid? The plots below try to capture a potential correlation. The plot on the left shows per month, the total number of passengers traveling to a particular state, where each state is color-coded. The plot on the right shows the state cumulative Covid cases per month, where each state is color-coded. The comparison between the two plots is inconclusive however. For all states, the cumulative number of Covid cases increases per month exponentially adjusted by population. It is difficult to state whether this is directly correlated with air travel, because in April when flights are essentially at a standstill, all states having continued rising Covid cases. However, it can be seen that as air travel begins to pick back up in the later part of 2020, the cumulative Covid cases per state begins to rise more exponentially. Based on this airline data, it is difficult to correlate whether airline travel to specific states caused these specific Covid spikes in these regions, as it could be due to lack of data at the beginning of the pandemic, lack of testing, or even other policies such as social distancing and mask wearing.


```{r,fig.width=10,fig.height=4}
# Aggregated State Airline Data - passengers per month by state
state_data <- air_data %>%
  group_by(DEST_STATE_ABR, MONTH, YEAR) %>%
  summarise(Total_Passengers = sum(PASSENGERS), Total_Departures = sum(DEPARTURES_PERFORMED)) %>%
  mutate(MONTH = month.abb[as.numeric(gsub("M","",MONTH))])
colnames(state_data)[1] <- "State"

# monthly cumulative covid case data per month by state

covid_state_cum_monthly$YEAR <- format(covid_state_cum_monthly$month, format="%Y")
covid_state_cum_monthly$MONTH <- format(covid_state_cum_monthly$month, format="%b")

# merge two data sets
state_air_covid <- merge(state_data, covid_state_cum_monthly, by=c("State","YEAR","MONTH")) %>%
mutate(MONTH = factor(MONTH, levels=c("Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct","Nov","Dec")))

# Plot Side by side Passengers per State and Total Cases per State by Month
plt1 <- state_air_covid %>%
ggplot(aes(MONTH, Total_Passengers, color=State)) + 
  geom_point() + 
  ylab("Total Passengers") +
  scale_y_continuous(labels = scales::comma) + 
  theme(legend.position = "none") +
  ggtitle("State Total Inbound Passengers")

plt2 <- state_air_covid %>%
ggplot(aes(MONTH, positive_monthly, color=State)) + 
  geom_point() + 
  ylab("Total Covid Cases") +
  scale_y_continuous(labels = scales::comma) + 
  theme(legend.position = "none") +
  ggtitle("State Total Cases")

gridExtra::grid.arrange(plt1,plt2,ncol=2,nrow=1)
```

Taking a look at specific states might give us more information. Florida and New York are two states which have similar populations, and handled the pandemic in completely different ways. Florida had an open door policy and welcomed visitors from all states during the height of the pandemic, while the densely populated New York followed mask mandates and social distancing. Can we see this difference in policy visually? 

The plots below isolate the two plots shown above for these two states. It is clear from the plot on the left, after the initial April shut down of flights, the total passenger travel to Florida as a destination significantly increased in comparison to flights to New York. This is due to people in the East Coast (New York, Chicago, Georgia etc.) escaping the winter for the hot beach weather in Florida. Taking a look at the cumulative Covid cases on the right plot, initially New York was significantly harder hit in the months of April to June in comparison to Florida. However, from July onward, Florida Covid cases spiked upwards and overtook New York. We can state that the continuous increase in traffic to Florida may have played a role in the increasing Covid cases, however, as discussed above, it could be just due to a lack of social distancing and mask wearing that led to this rise. Travel to New York was consistently down from April to December, and yet the cumulative cases continued to rise in the later months, to a similar number as Florida. 


```{r, fig.width=10,fig.height=4}
# Two similar populations Comparison (NY and FL)
states <- c("NY", "FL")
compare_air_covid <- state_air_covid %>%
  filter(State %in% states)

plt1 <- ggplot(compare_air_covid, aes(MONTH,Total_Passengers, group=1,colour=State)) + 
  geom_point() + 
  ylab("Total Passengers") +
  geom_point() + 
  scale_y_continuous(labels = scales::comma)+
  scale_color_colorblind() +
  ggtitle("State Total Inbound Passengers")
  
plt2 <- ggplot(compare_air_covid, aes(MONTH,positive_monthly, group=1,colour=State)) + 
  geom_point() + 
  ylab("Total Covid Cases") +
  geom_point() + 
  scale_y_continuous(labels = scales::comma)+
  scale_color_colorblind() +
  ggtitle("State Total Cases")

gridExtra::grid.arrange(plt1,plt2,ncol=2,nrow=1)


```


The only way to conclude the correlation between flight traffic and Covid spikes is with daily flight data. The data from the Bureau of Transportation Statistics aggregates flight data by month. If we had access to daily flight data for 2020, that would give us the ability to compare with daily Covid cases and see if increased air travel during a specific time (ex. spring break) accounted for a Covid spike in the following two weeks. However, this flight data is not accessible, maybe only through commercial resources, and the amount of flights daily in America, let alone in a year, would be a huge dataset that would be difficult to pre-process.
