==> 07-conclusion.Rmd <==
# Conclusion

This project was a fascinating journey through the data of a nation, and a pandemic. Our findings suggest that the mobility of people may not have been such a strong influencing factor in the spread of COVID-19. Any inconclusive results of this project are due to the complexity of this virus and civilization. A national, let alone global, pandemic is an incredibly complex problem. On top of that, we are dealing with the mobility of millions of people. Understanding how Covid-19 spread in the US was an interesting challenge, but with the data we were able to source, there were just too many external factors associated with pinpointing specific causes: policy variation such as mask mandates and lockdowns across states, data reporting variations across states, and the simple fact that this virus has never been seen before globally and understanding it’s transmission is still being researched today. A future direction to understanding the spread of Covid through mobility, would be comparing the US data we observed to countries that had stricter lockdown measures such as England, New Zealand or Taiwan. This project gave us an opportunity to explore and combine different datasets and observe whatever correlations we could find within the data through visualization, even if that meant inconclusive results. We thoroughly enjoyed the process and hope to further this Covid study as more data is available in 2021 with vaccinations!
